+++ sync-for-payables/RELEASES.md
@@ -97,3 +97,13 @@ Based on:
 - [csharp v2.2.0] sync-for-payables
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Sync.Payables/2.2.0 - sync-for-payables
+
+## 2023-10-02 00:22:35
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.3.0] sync-for-payables
+### Releases
+- [NuGet v2.3.0] https://www.nuget.org/packages/Codat.Sync.Payables/2.3.0 - sync-for-payables